Create npm package and webpack config boilerplate
Summary:
This took way to long to figure out,
because the Babel 5 to Babel 6 upgrade changed some things
that I didn't expect to change.
It seems to be working now, though.

Test Plan:
Execute `npm start` and visit the page in a web browser.
See that React loads and a header is rendered.