Implement findChordRootOffset
Summary:
This is at least somewhat of a shot in the dark.
I'm assuming that we define root-finding recursively:
that is, the root of a binary chord is just the root of the interval,
and for any other chord we find the roots of adjacent notes
and recurse on those results themselves (with one fewer pitch).
I'm not sure I always agree with the results this gives out, though;
for example, C-F-Ab is resolved to a D-flat.
So we'll see.

Test Plan:
Unit tests included (for this specification, anyway).
Run `npm test`.