Add a rationalizer configuration component
Summary:
This component provides a bunch of text boxes
for specifying the acoustic ratios for the simple semitone differences.
It defaults to the values of the canonical rationalizer.

The editing mode is currently entirely string-based:
you make changes, then we try to parse those changes with a regex;
if they work, the result is saved as a Rational.
(The underlying value is always stored as a Rational, never a string.)
This seems to actually work out pretty well;
it's probably going to prove a lot less cumbersome than, say,
two separate text boxes (one for the numerator, one for denominator).
The simplification mechanism---namely, you can enter whatever you want
but as soon as you blur the fraction will be simplified
(e.g., "4:2" becomes "2:1" when you exit the text field)---
is quite nice.

Doesn't include nice error handling,
but I don't think it really needs to:
you'll know if you divide by zero or something stupid,
and if you enter something that doesn't match the regex
you probably just have a typo (`:` to `;`, maybe?)
and should just try again.

Screenshot:
![Screenshot](http://i.imgur.com/VXTV22M.png)