Remove bolding from changed mappings
Summary:
Previously, we made it so that if you change an acoustic ratio
that cell will be marked in bold.
That came off a bit too strong;
hiding the reset buttons for unchanged ratios works much better.
So we keep the latter and nuke the former.

Test Plan:
Change some ratios. Note that they don't turn bold.