Put chord notes on separate lines in tree explorer
Summary:
Previously, we wrote something like "C4, E4, G4."
But it makes more sense to have those stacked vertically
and in the actual chord order.
So we do that.

Test Plan:
Visit /#/tree and make sure that the note names are on separate lines
and the chord is in the right order;
that is, if the chord is "C4, E4, G4," then C4 should be on the bottom.