Show the root of trichords in the tree explorer
Summary:
This currently always uses the canonical rationalizer,
because we don't have a way to access the user's custom data.
We should add that functionality, and I see two ways to do so:
 (1) grab it from the LocalStore,
     and probably make the RationalizerConfig a top-level component
     and include it on the tree explorer page;
 (2) introduce a Redux store,
     and move the logic for talking to the LocalStore
     into some sort of subscription.
The second option is probably a bit "purer,"
but really the first doesn't sound so bad.
We'll see how the wind blows.

Test Plan:
Visit /#/tree and verify that
 (1) the roots of chords appear;
 (2) the roots appear in blue; and
 (3) the roots are correct
     (i.e., consistent with the root calculator
     when using the canonical rationalizer);