Add error handling for infinite roots
Summary:
When the acoustic ratios have really big numerators and denominators,
the exponentiation by scalars can cause an overflow to `Infinity`.
This causes the `Rational` constructor to throw an error.
We now check for those and alert the user accordingly.

This is a quickfix; soon, we'll change those functions
from returning a `Rational` and sometimes throwing
to returning an `Either String Rational`, as should be the case.

Test Plan:
Try `-3473, 9761, 14945`.