Make harmonic error checking properly functional
Summary:
In the previous commit, we fixed some issues with running into Infinity.
But we did so with `try`-`catch`, and that was a bit not great.
Now we do so at the source:
select methods that might fail return an `Either String a`
instead of maybe returning an `a` and maybe throwing.

Test Plan:
Unit tests included for the core. Run `npm test`.
Manually test some core GUI core functionality.
An error case to try is `-3473, 9761, 14945`;
this should work fine in both the root calculator and the tree explorer.