Split semitone counts onto new lines in tree view
Summary:
That is, put the `[m][n]` on two lines instead of one.
This makes the boxes smaller,
which is especially nice when combined with "wide" mode
and octave numbers disabled.

Consequently, we also increase the tree depth limit to eight levels.
That's too big for my (1080p) monitor at normal 100% zoom,
but there's no reason to limit anyone who could take advantage of it!