Add engraved notes to the tree view on hover
Summary:
Good news: this seems to work in simple cases!
Mixed news: it revealed why VexFlow was being so tricky sometimes.
In particular, the bounding box returned by the formatter
doesn't include the staff, only the notes;
so if you have a bunch of notes that are all below the staff
the bounding box will be way too small.
We'll have to find a way around that…or just deal with it? ¯\_(ツ)_/¯

Test Plan:
Visit `/#/tree` and hover things!