Fix engraving padding problem!
Summary:
The issue, as mentioned in 66425a52a653891ad107e0757d4e2b7581328de0,
was that the bounding box returned by `FormatAndDraw`
only accounted for the noteheads themselves,
so if they all appeared below the staff but were spaced closely together
the height of the bounding box would be grossly underestimated.
This obviously made the spacing all wrong.

The solution is to use the built-in `getBBox` method of SVG elements.
I was hesitant to use this previously because
for it to work properly (i.e., not return `(0, 0, 0, 0)`),
the SVG node must be already mounted in the DOM,
so I was concerned that there would be some "flickering"
between the time we mount the SVG and the time we set all the margins.
But that doesn't appear to be the case at all.
Presumably this is due to the "JavaScript blocks the DOM" effect:
the browser needs the SVG to be in the DOM
so that it can compute the styles (e.g., font sizes)
and any transformations (e.g., scales, skews, translations),
but it doesn't actually render it to the user until the JS is all done.

Yay; it's the best of both worlds!

Test Plan:
Set the tree explorer root to `-60, -55, -50`
and note that it renders just dandy.