Add minimum/maximum limits UI (not yet functional)
Summary:
Wow. This is a lot of state.
It's also a good deal of abstraction.
I think it's worth it so far, but the whole thing is a bit unsettling.
We'll keep an eye on this.

I played around with using nullable numbers
instead of having separate fields for "value" and "enabled,"
but that incurs the issue of what to do
when the user enables the limit after it's been disabled:
do you pick a random number for the limit?
It'd be nice if we remembered what the previous setting was.
So we do.