Implement a Redux-style reducer for the app state
Summary:
The application state now contains all configuration,
like the acoustic ratios and the tree view options,
but not any of the transient data,
like what chord is currently being analyzed.
I figure we'll start with this and integrate it,
then see whether we want to add more things.

The benefit of this is that we'll be able
to serialize and deserialize the whole application state at once,
which includes saving it in the LocalStore.

Test Plan:
Unit tests included. Run `npm test`.