Write the Redux store to the LocalStore on change
Summary:
This is a one-way binding.
For it to be complete, we need to rehydrate the state on startup.
We haven't yet done that.

Test Plan:
Make a change that dispatches some Redux action
(e.g., change one of the acoustic ratios),
then execute `LocalStore.get('redux_state')` in the JavaScript console.
You should see an object with the updated world state.