Integrate the Redux rehydration with LocalStore
Summary:
All the pieces are now in place!
We can serialize the entire application state whenever we make a change,
and we can rehydrate it when we launch the application.

Test Plan:
Load the application and change one of the acoustic ratios.
Refresh and note that the change persists.
Then, run `LocalStore.del('redux_store')` in the JS console.
Refresh again.
The store should have returned to its original state,
and—most importantly—the application should not have crashed!