Fix mouse capture issue on really tall engravings
Summary:
Previously, if you had a really tall engraving,
we would position it correctly,
but it would still capture pointer events
because the DOM thought it was lower than it actually was:
![Demo image](https://cloud.githubusercontent.com/assets/4317806/11758618/103484ce-a022-11e5-8c68-e9afd8ce790f.png)
We fix that by just disallowing pointer events.
(You can't click them, anyway,
because the hover overlay would disappear
as soon as you moved your mouse off the button;
plus, nothing would happen if you clicked them,
so why are you even trying?)

Test Plan:
Replicate the above scenario and note that you can click on the button.