Increase engraving z-index to move above forms
Summary:
We had a z-index of 1 on the engravings
to put them above the buttons in the chord tree.
This was all fine and dandy until we added the Bootstrap `input` boxes,
which apparently have a z-index of 2.
We can fix that just by increasing this z-index one more time!

Test Plan:
Load a really tall chord and note that it appears in front of
the inputs used to set the limits.