Use the Redux state rationalizer in the tree view
Summary:
This involves hooking up the root `TreeExplorer` to Redux,
and just passing down the rationalizer a few levels!

One subtlety is that it introduces a new kind of error:
the user's allowed to enter zero ratios,
which may carry over into the tree view,
so we have to handle those.
Presently, we handle those, although not very *well*:
the error message is lumped into the existing error;
that is, the error will be caught and handled,
but the message will be completely wrong.
So there's a TODO to fix that.