Add a RationalizerConfig to the tree view
Summary:
Nominally, this is only shown when roots are shown,
but the settings to control that aren't yet hooked up to Redux,
so for the time being it'll always be visible.
But the ratios themselves are hooked up to Redux,
so you can edit them and see the state change.