Move limit settings into Redux
Summary:
This is the last step in the great migration! Everything's stateless!
This one is particularly satisfying because
I was never satisfied with the way the limit settings were designed;
the whole existence of the `limitHandlers` object made me sad.
Now the actions abstract over all that, which is neat-o.