Force tree nodes to be aligned in a grid
Summary:
This makes me undescribably happy.
Previously, there was this annoying behavior where
the nodes of the tree could have slightly different widths,
just because some of the numbers were wider than others
or some of them had minus signs or accidentals while others didn't.
For example, a node with a "G♯−8" is a lot wider than a "D1".
This caused the nodes to be slightly offset from each other,
which completely ruined the tree structure.

Turns out the fix is as simple as a `width` property!
Wonderful!