Make the root field of the trichord tree editable
Summary:
This is a great example of the 80–20 rule!

The basic approach to the problem is rather simple:
just make the root node not a button,
because clicking it wouldn't do anything anyway,
and swap out the notes for text fields.
Then use normal change handlers to update the state.

Here's the catch:
what happens when the chord is C4, E4, G4
and the user changes the middle note to an E5?
This has never been a problem before, because
the entire chord has always been edited within one text field.
But that's no longer the case:
now, the new values of the text fields
should be C4, G4, and E5, respectively.
That is, we need to swap the G4 and E5 fields
in addition to updating the edited field's value.

The implementation of this is a bit tricky,
but I've added a reasonably thorough explanation
in the doc comment for SingleNoteInput,
so I'll leave you to browse that.

One possible area for improvement:
the whole business with "is the input array sorted? which way?"
and "am I indexing into the ascending or descending array?"
is needlessly complicated.
Perhaps this could be changed by
creating all the elements and handling all the events
in ascending order,
and just rendering them in descending order?

Test Plan:
Test cases include:
  * make "normal" changes (e.g., C4, E4, G4 to C4, F4, G4);
  * change the middle note to be above the top note;
  * change the bottom note to be above the top note;
  * do the inverses of those two (i.e., middle or top below bottom);
  * change one of the notes to be identical to one of the others.