Simplify ChordEngraving positioning logic
Summary:
We had a lot of crazy heuristics in there
that seemed to work at the time but actually don't make much sense.
The need for them is gone now that we're using the proper bounding box
(see 50727d099de1c95025ac3ec67e5e320aed4e3519 for details),
and everything actually looks *better* this way!

Test Plan:
Set the root chord to C4, G4, B♭4, and set the depth to 5.
Hover over each node and make sure it looks right.
This particular test case gives a good range of inputs:
the root is entirely within the staff, but some chords are really wide.

Also check a chord like C1, D1, E1, which is entirely below the staff,
and one like C8, E8, E8, which is entirely above the staff.