Include octave names for default display text
Summary:
When the user presses UpArrow or DownArrow on an input field,
we can easily update the actual pitch in the chord,
but we also need to update the display text for the input field.
Instead of trying to guess the user's format and remain consistent
(like "you wrote 'g#4' and pressed UpArrow so I'll give you 'a4'"),
which seems like a bunch of trouble for no apparent gain,
we just generate the canonical display string.
This is fine.

But previously we were only including the octave in the generated string
if the user had indicated that the octaves should be shown.
This is inconsistent with the rest of the input field behavior,
and with the underlying reason:
we always show the octave once you click the text box
because the user has to enter the octave when they type in a note.

So, in short, we just want to always show the octaves in the input.