Add some math notes for a new infinite canvas
Summary:
Last time, I kind of hacked around to try to get things to work.
That kind of worked, but not very well.
So now I went through and actually did the math,
and I'll try to re-implement it with all that in mind.

I'm keeping the old file for reference for now.