Add nodes in the right positions
Summary:
This is a proof-of-concept for mapping
absolute coordinates (in extended canvas space) to viewport coordinates.
It seems to work!

Test Plan:
Pan around the tree.
Note that the zoom adjusts appropriately;
i.e., if you move some node to the top middle,
it looks just like that original position.
Note also that you can't scroll out of bounds horizontally.