Only render nodes in view
Summary:
Previously, we were just rendering a bunch of static rows,
and if you went out of bounds, too bad.
Now, we can determine which rows are visible
just based on your `y` position,
and which nodes in the row are visible
based on your `x` position and a few other calculations.
So viewport rendering is now constant with respect to position,
which means you can dive as deep into the tree as you like!
(Well, until you get about 1024 levels deep and overflow Number.)