Move instance properties to the constructor
Summary:
The rule of thumb is that each instance property
should be initialized to some value in the constructor.
This way, it's easy to see what all the instance properties are.
Sometimes, we can move the actual value into the constructor,
as with _resizeListener.
Other times, we just initialize it with a null value,
as with _keyInterval.