Create a more declarative settings table framework
Summary:
Previously, we were using a bunch of ad-hoc styles for the settings.
Now we've extracted those to dedicated semantic classes.
This way, they can also be shared.

As a consequence, ViewOptions is now as simple as can be:
it's a class with just a render method,
and a render method with just a return statement!