Simplify the padding logic by just painting more
Summary:
Previously, we were assuming a fixed horizontal and vertical padding,
which would presumably be constant or easily computable,
and using that to precisely bound the rows and columns to paint.
But it looks like determining that isn't so easy after all,
because it depends on the font size and maybe some context metrics,
so really it's just a lot easier to do more work.
Specifically, render an extra row and an extra column on each side.