Fix a complex scale bug
Summary:
Again, this had to do with painting things outside of the screen.
If the `y` position was greater than the canvas
then the scale would be the square root of a negative number,
and that made some things sad.

(I guess really the bug was purely imaginary,
but that sounds less humorous and more confusing.)