Make drag behavior non-accumulative
Summary:
Because the scaling factor is non-linear,
it used to be the case that
dragging up and to the right in one motion
was different from dragging up and then dragging to the right
(releasing the mouse in between).
More generally, the drag behavior is non-commutative,
but we were basically assuming commutativity to simplify the motion.
This was a bit confusing.
With the new behavior, deltas are computed at every mouseMove event.