Use actual note values in infinite tree!
Summary:
This was more difficult than expected
because it turns out that my previous approach was losing information!
If you only track the semitone differences,
you lose track of what the bass of the chord is
(e.g., if you outfold up then outfold up again,
none of the notes in the chord is on the original bass.)
So I had to rewrite the underlying TreeSpace function.
(It's actually shorter now!)
From there, it was easy!