Add a background color to the infinite tree cells
Summary:
This is a bit tricky because
we calculate the box metrics while we're painting the text,
so if we then fill it we cover up all the text.
We "solve" this by just repainting it.

The other obvious solution is to calculate the metrics
without drawing the text initially.
That's certainly "nicer,"
but I doubt it would have any real performance benefit
so I'm just not doing it for now.
(Low-hanging fruit for later.)