Extract `withinLimits` to new DisplayUtils module
Summary:
Also moving to the "don't maintain a separate export dictionary" style,
where you have to `import * as Foo from './Foo'` if you want everything,
because that seems cleaner in the long run.
I've learned to avoid *-imports because they tend to pollute namespaces,
but that's not the case in JavaScript (unlike Java or Python)
because you have to assign them to an object.

Test Plan:
Unit tests included. Run `npm test`.