Change infinite canvas coordinate system
Summary:
This changes the idealized system to work in more idealized units.
In particular, the scaling factor is canvas-independent;
a change of 1 y-unit represents one level,
and the entire horizontal axis ranges from -0.5 to +0.5.

The main benefit of this change is that
when you change any of the view options,
you no longer randomly lose your place in the tree!