Use a more accurate 1.2 bold-weight heuristic
Summary:
Bold characters really aren't 50% wider;
this was causing a string like "B♭6" to be chosen over "E♭15",
which in turn caused the bounding box to be too small.
Now it basically just means
"choose bold if there's a tie or the strings are *really* long,"
which is what it should mean.