Implement `getNodeCanvasCoordinates` on CanvasCore
Summary:
This uses a slightly different implementation from the existing canvas:
the nodes' tops are no longer centered on the rows,
but instead are aligned to the top of the row with a bit of padding.
This probably makes more sense;
I'll change the existing canvas implementation next.

Test Plan:
Unit tests included. Run `npm test`.