Make `makeSafeRow` a number, not constant function
Summary:
There's really no point in having it be a constant function.
(In fact, the Haskell types
for a constant function and for a "non-function" value
are identical: they're both just the "non-function" value type,
because the notion of "constant function" is isomorphic to that of "value",
because 1 × τ ≅ τ for all types τ.)

Test Plan:
Existing unit tests suffice. Run `npm test`.