Begin merging `InfiniteCanvas` and `CoreCanvas`
Summary:
This replaces the outer `_draw` function of `InfiniteCanvas`
to simply delegate to `CoreCanvas` functions.
We're not yet using `CoreCanvas` for the event handling
(specifically, the panning),
so we currently have to manufacture our state anew every frame.

Test Plan:
See that the canvas behaves the same, without any observable changes,
even when you scroll to the bottom!