Deduplicate canvasHeight property
Summary:
Previously, we had a canvasDimensions property on the core state
that stored both width and height.
But we always have height accessible as `state.viewOptions.infiniteHeight`,
so it doesn't make sense to let these get out of sync.
Happily, this only requires one change in the source,
because all the accesses are indirect except for `getRowDimensions`.

Test Plan:
Existing unit tests suffice. Run `npm test`.