Begin implementation of StackedNoteInput
Summary:
This is meant to be a functional, pure, stateless replacement
for the multiple `SingleNoteInput` components currently in use.
The wiring between those existing components
consists of a bunch of stateful manipulations
(like the `takeStateFrom` method),
which further span multiple components in the hierarchy
(so the mutation isn't scoped within a single file).
This is sad, and makes testing hard, and should be fixed.

This commit adds a partial implementation of the existing behavior.
Specifically, it can render multiple inputs and modify them properly,
and it supports moving one note above another note
(which is the cause of most of the statefulness
in the previous implementation).
I think the only thing that's missing are the arrow key bindings
(Up, Down, Shift-Up, and Shift-Down)
for nudging notes.

This includes comprehensive unit tests
of both the DOM output and simulated user interactions.

Test Plan:
Unit tests included. Run `npm test`.