Update to Babel 7 and Webpack 4
Test Plan:
Run `npm start` and navigate to the dev server in a browser, then modify
`src/components/App.jsx`, and note that the page reloads with the new
content. Then, run `npm run coverage` and note that all tests run (and
pass) with a coverage reported emitted to `./coverage`.