Switch from @babel/polyfill to bare core-js + regenerator-runtime
Summary:
The `@babel/polyfill` package is deprecated, and this is the recommended
path forward. This change also lets us pin to `core-js@3.0.1`, which is
the latest version without bundled adware. We remove `@babel/node` for
the same reason (it was not actively used).

Test Plan:
Add a generator to the bundle…

```javascript
diff --git a/src/index.jsx b/src/index.jsx
index e28b919..10ccf12 100644
--- a/src/index.jsx
+++ b/src/index.jsx
@@ -22,3 +22,9 @@ const component = <Provider store={store}>
 </Provider>;

 ReactDOM.render(component, document.getElementById('app'));
+
+window.testtest = function*() {
+    yield 1;
+    yield 2;
+    return 3;
+}
```

…then verify `String(window.testtest)` that the function is compiled
with `regenerator-runtime` and via `[...testtest()]` that the function
works properly.

Running `npm run coverage` still works.